Partial fix for kanban node dropdown menus
Problem:
Since the kanban column requires overflow: auto, the dropdown will allways be
clipped. Proper solution would be to move the dropdown to a different position
in the DOM so it is not a child of any element that has clipping active.

The solution via (the now removed) overrideOverflowCutoff method did not work
in firefox & android.